+++ src-tauri/src/database/spaces.rs
@@ -10,7 +10,7 @@ pub async fn create_space(
     name: &str,
     llm_config: LLMConfig,
     embedding_config: EmbeddingConfig,
-) -> Result<i64> {
+) -> Result<i32> {
     let llm_config_json = Json(llm_config);
     let embedding_config_json = Json(embedding_config);
     let record = sqlx::query!(
@@ -26,7 +26,7 @@ pub async fn create_space(
     .fetch_one(pool)
     .await?;
 
-    Ok(record.id)
+    Ok(record.id as i32)
 }
 
 pub async fn get_all_spaces(pool: &DbPool) -> Result<Vec<Space>> {
@@ -37,7 +37,7 @@ pub async fn get_all_spaces(pool: &DbPool) -> Result<Vec<Space>> {
     Ok(res)
 }
 
-pub async fn add_root(pool: &DbPool, space_id: i64, path: &str) -> Result<i64> {
+pub async fn add_root(pool: &DbPool, space_id: i32, path: &str) -> Result<i32> {
     let record = sqlx::query!(
         "INSERT INTO indexed_roots (space_id, path) VALUES (?, ?) RETURNING id",
         space_id,
@@ -46,5 +46,5 @@ pub async fn add_root(pool: &DbPool, space_id: i64, path: &str) -> Result<i64> {
     .fetch_one(pool)
     .await?;
 
-    Ok(record.id.context("failed to retrieve inserted root ID")?)
+    Ok(record.id.context("failed to retrieve inserted root ID")? as i32)
 }
